% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coord_proj.r
\name{coord_proj}
\alias{coord_proj}
\title{Like \code{coord_map} only better :-)}
\usage{

  coord_proj(proj = "+proj=robin +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs",
  inverse = FALSE, degrees = TRUE, ellps.default = "sphere",
  xlim = NULL, ylim = NULL)
}
\arguments{
\item{proj}{projection definition}

\item{inverse}{if \code{TRUE} inverse projection is performed (from a
cartographic projection into lat/long), otherwise projects from
lat/long into a cartographic projection.}

\item{degrees}{if \code{TRUE} then the lat/long data is assumed to be in
degrees, otherwise in radians}

\item{ellps.default}{default ellipsoid that will be added if no datum or
ellipsoid parameter is specified in proj. Older versions of PROJ.4
didn't require a datum (and used sphere by default), but 4.5.0 and
higher always require a datum or an ellipsoid. Set to  \code{NA} if no
datum should be added to proj (e.g. if you specify an ellipsoid
directly).}

\item{xlim}{manually specific x limits (in degrees of longitude)}

\item{ylim}{manually specific y limits (in degrees of latitude)}
}
\description{
The representation of a portion of the earth, which is approximately
spherical, onto a flat 2D plane requires a projection. This is what
\code{coord_proj} does, using the \link[proj4]{project()} function from
the \code{proj4} package.
}

