% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_table.r
\name{geom_table}
\alias{geom_table}
\title{Add a table to a ggplot2 plot}
\usage{
geom_table(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Add a table to a ggplot2 plot
}
\author{
Jan Schulz
}

