% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_xspline2.r
\name{geom_xspline2}
\alias{geom_xspline2}
\title{xsplines}
\usage{
geom_xspline2(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{stuff}
}
\value{
creates a spline curve
}
\description{
Xspline
}
\author{
Ben Bolker
}

