% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_bkde.r, R/geom_bkde2d.r, R/geom_xspline.r
\docType{data}
\name{GeomBkde}
\alias{GeomBkde}
\alias{GeomBkde2d}
\alias{GeomXspline}
\alias{StatBkde}
\alias{StatBkde2d}
\alias{StatXspline}
\title{GeomXspline}
\description{
GeomXspline
}
\keyword{datasets}

