% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalt-package.r
\docType{package}
\name{ggalt-package}
\alias{ggalt}
\alias{ggalt-package}
\title{ggalt: Extra Coordinate Systems, 'Geoms', Statistical Transformations, Scales and Fonts for 'ggplot2'}
\description{
A compendium of new geometries, coordinate systems, statistical transformations, scales and fonts for 'ggplot2', including splines, 1d and 2d densities, univariate average shifted histograms, format functions for "bytes", a stat_stepribbon() function, and the 'StateFace' open source font 'ProPublica'. Further new functionality includes lollipop charts, dumbbell charts, the ability to encircle points and coordinate-system-based text annotations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/ggalt}
  \item Report bugs at \url{https://github.com/hrbrmstr/ggalt/issues}
}

}
\author{
\strong{Maintainer}: Bob Rudis \email{bob@rud.is} (\href{https://orcid.org/0000-0001-5670-2640}{ORCID})

Authors:
\itemize{
  \item Ben Bolker (Encircling & additional splines) [contributor]
  \item Jan Schulz (Annotations) [contributor]
  \item Aditya Kothari (Core functionality of horizon plots) [contributor]
  \item Jonathan Sidi (Annotation ticks) [contributor]
}

Other contributors:
\itemize{
  \item Ben Marwick (General codebase cleanup) [contributor]
  \item Rosen Matev (Original annotate_textp implementation on stackoverflow) [contributor]
  \item ProPublica (StateFace font) [data contributor]
  \item Ather (Core functionality of horizon plots) [data contributor]
  \item Tarcisio Fedrizzi (Bytes formatter) [contributor]
}

}
\keyword{internal}
