% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_ticks.r, R/geom_cartogram.r
\docType{data}
\name{GeomTicks}
\alias{GeomTicks}
\alias{GeomCartogram}
\title{Base ggproto classes for ggplot2}
\description{
If you are creating a new geom, stat, position, or scale in another package,
you'll need to extend from ggplot2::Geom, ggplot2::Stat, ggplot2::Position, or ggplot2::Scale.

Geom Cartogram
}
\seealso{
\code{\link[ggplot2]{ggplot2-ggproto}}
}
\keyword{datasets}
