% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.r
\name{list_avatars}
\alias{list_avatars}
\title{List Pokemon avatar names}
\usage{
list_avatars()
}
\description{
Data taken from the hard work by the \href{http://www.pokegraphs.com/}{folks here}.
}
\note{
Warning: huge! list
}

