% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateface.r
\name{load_stateface}
\alias{load_stateface}
\title{Load stateface font}
\usage{
load_stateface()
}
\description{
Makes the ProPublica StateFace font available to PDF, PostScript,
et. al. devices.
}
\seealso{
Other StateFace operations: \code{\link{geom_stateface}},
  \code{\link{show_stateface}}
}
