% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom2plotly.r
\name{plotly_helpers}
\alias{plotly_helpers}
\alias{to_basic.GeomXspline}
\alias{to_basic.GeomBkde2d}
\alias{to_basic.GeomStateface}
\title{Plotly helpers}
\usage{
to_basic.GeomXspline(data, prestats_data, layout, params, p, ...)

to_basic.GeomBkde2d(data, prestats_data, layout, params, p, ...)

to_basic.GeomStateface(data, prestats_data, layout, params, p, ...)
}
\arguments{
\item{data, prestats_data, layout, params, p, ...}{plotly interface parameters}
}
\description{
Helper functions to make it easier to automatically create plotly charts
}
\keyword{internal}
