% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-dodgev.R
\docType{data}
\name{position_dodgev}
\alias{position_dodgev}
\alias{PositionDodgev}
\alias{pos_dodgev}
\alias{collidev}
\title{Vertically dodge position}
\usage{
position_dodgev(height = NULL)

pos_dodgev(df, height)

collidev(data, height = NULL, name, strategy, ..., check.height = TRUE,
  reverse = FALSE)
}
\description{
Vertically dodge position
}
\note{
position-dodgev(): unmodified from lionel-/ggstance/R/position-dodgev.R 73f521384ae8ea277db5f7d5a2854004aa18f947

PositionDodgev(): based on from lionel-/ggstance/R/position-dodgev.R 73f521384ae8ea277db5f7d5a2854004aa18f947

pos_dodgev(): unmodified from lionel-/ggstance/R/position-dodgev.R 73f521384ae8ea277db5f7d5a2854004aa18f947

collidev(): based on lionel-/ggstance/R/position.R 73f521384ae8ea277db5f7d5a2854004aa18f947
}
\author{
@ggstance authors

@ggstance authors

@ggstance authors

@ggstance authors
}
\keyword{datasets}
