% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-jitter-on-equal.R
\name{position_jitter_on_equal}
\alias{position_jitter_on_equal}
\title{Jitter points to avoid overplotting}
\usage{
position_jitter_on_equal(amount = 0)
}
\arguments{
\item{amount}{Amount of vertical jitter (0..1)}
}
\description{
Designed primarily for [geom_dumbbell()] to jitter the second point vertically
when both x values are equal.
}
\concept{position adjustments}
