% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.r
\name{scale_colour_pokemon}
\alias{scale_color_pokemon}
\alias{scale_colour_pokemon}
\alias{scale_fill_pokemon}
\title{Pokemon color scales}
\usage{
scale_colour_pokemon(avatar = "bulbasaur", ...)

scale_color_pokemon(avatar = "bulbasaur", ...)

scale_fill_pokemon(avatar = FALSE, ...)
}
\arguments{
\item{avatar}{avatar name. Use \code{list_avatars()} to see them all!}

\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
Color scales using the colors in the Pokemon characters
}
\note{
Pokémon & Pokémon character names are trademarks of Nintendo.
}
\references{
\href{http://www.pokegraphs.com/}{Original JSON color list}.
}

