% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateface.r
\name{show_stateface}
\alias{show_stateface}
\title{Show location of StateFace font}
\usage{
show_stateface()
}
\description{
Displays the path to the StateFace font. For the font to work
in the on-screen plot device for ggplot2, you need to install
the font on your system
}
\seealso{
Other StateFace operations: \code{\link{geom_stateface}},
  \code{\link{load_stateface}}
}
