% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_bkde2d.r
\name{stat_bkde2d}
\alias{stat_bkde2d}
\title{2D density}
\usage{
stat_bkde2d(mapping = NULL, data = NULL, geom = "density2d",
  position = "identity", contour = TRUE, bandwidth, grid_size = c(51, 51),
  range.x = NULL, truncate = TRUE, show.legend = NA, inherit.aes = TRUE,
  ...)
}
\arguments{
\item{contour}{If \code{TRUE}, contour the results of the 2d density
estimation}
}
\description{
2D density
}
\section{Computed variables}{

Same as \code{\link{stat_contour}}
}

