% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dumbbell.R
\name{geom_dumbbell}
\alias{geom_dumbbell}
\title{Dumbbell charts}
\usage{
geom_dumbbell(
  mapping = NULL,
  data = NULL,
  ...,
  colour_x = NULL,
  size_x = NULL,
  shape_x = NULL,
  fill_x = NULL,
  colour_xend = NULL,
  size_xend = NULL,
  shape_xend = NULL,
  fill_xend = NULL,
  dot_guide = FALSE,
  dot_guide_size = NULL,
  dot_guide_colour = NULL,
  jitter_on_equal_values = 0,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  position = "identity"
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{colour_x, colour_xend}{the colour of x/xend; defaults to colour aes}

\item{size_x, size_xend}{the size of the x/xend; defaults to 0.5}

\item{shape_x, shape_xend}{shape to use for x/xend; defaults to 19}

\item{fill_x, fill_xend}{fill for x/xend if the shape is hollow; defaults to fill aes}

\item{dot_guide}{if \code{TRUE}, a leading dotted line will be placed before the left-most dumbbell point}

\item{dot_guide_size, dot_guide_colour}{singe-value ahesthetics for \code{dot_guide}}

\item{jitter_on_equal_values}{numeric value [0..1], default is 0. If non-zero and if both
values are equal, jitter the points vertically by the amount specified so
they don't fully overlap.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}
}
\description{
The dumbbell geom is used to create dumbbell charts.
}
\details{
Dumbbell dot plots — dot plots with two or more series of data — are an
alternative to the clustered bar chart or slope graph.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "segment")}
}

\examples{
library(ggplot2)

data.frame(
  trt = factor(LETTERS[1:5], levels = LETTERS[5:1]),
  l = c(20, 40, 30, 30, 50),
  r = c(70, 50, 10, 60, 80)
) -> df

ggplot(df, aes(y=trt, x=l, xend=r)) +
  geom_dumbbell(size=3, color="#e3e2e1",
                colour_x = "#5b8124", colour_xend = "#bad744",
                dot_guide=TRUE, dot_guide_size=0.25) +
  labs(x=NULL, y=NULL, title="ggplot2 geom_dumbbell with dot guide") +
  theme_minimal() +
  theme(panel.grid.major.x=element_line(size=0.05))

ggplot(df, aes(y=trt, x=l, xend=r)) +
  geom_dumbbell(size=3, color="#e3e2e1",
                colour_x = "#5b8124", colour_xend = "#bad744", shape_x = 15,
                dot_guide=TRUE, dot_guide_size=0.25) +
  labs(x=NULL, y=NULL, title="ggplot2 geom_dumbbell with dot guide") +
  theme_minimal() +
  theme(panel.grid.major.x=element_line(size=0.05))

## with vertical dodging
df2 <- data.frame(trt = c(LETTERS[1:5], "D"),
                 l = c(20, 40, 10, 30, 50, 40),
                 r = c(70, 50, 30, 60, 80, 70))

ggplot(df2, aes(y=trt, x=l, xend=r)) +
  geom_dumbbell(size=3, color="#e3e2e1",
                colour_x = "#5b8124", colour_xend = "#bad744",
                dot_guide=TRUE, dot_guide_size=0.25,
                position=position_dodgev(height=0.4)) +
  labs(x=NULL, y=NULL, title="ggplot2 geom_dumbbell with dot guide") +
  theme_minimal() +
  theme(panel.grid.major.x=element_line(size=0.05))
}
