% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_proj.r, R/geom_ash.r, R/geom_bkde.r, R/geom_bkde2d.r, R/geom_xspline.r
\docType{data}
\name{CoordProj}
\alias{CoordProj}
\alias{GeomBkde}
\alias{GeomBkde2d}
\alias{GeomXspline}
\alias{StatAsh}
\alias{StatBkde}
\alias{StatBkde2d}
\alias{StatXspline}
\title{Geom Proto}
\description{
Geom Proto

Geom Proto

Geom Proto

Geom Proto

Geom Proto
}
\keyword{datasets}
\keyword{internal}

