% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_proj.r, R/geom_ash.r, R/geom_bkde.r,
%   R/geom_bkde2d.r, R/geom_dumbbell.R, R/geom_encircle.r, R/geom_lollipop.r,
%   R/geom_xspline.r, R/geom_xspline2.r, R/stat-stepribbon.r, R/stateface.r
\docType{data}
\name{CoordProj}
\alias{CoordProj}
\alias{StatAsh}
\alias{GeomBkde}
\alias{StatBkde}
\alias{GeomBkde2d}
\alias{StatBkde2d}
\alias{GeomDumbbell}
\alias{GeomEncircle}
\alias{GeomLollipop}
\alias{GeomXspline}
\alias{StatXspline}
\alias{GeomXSpline2}
\alias{StatStepribbon}
\alias{GeomStateface}
\title{Geom Proto}
\description{
Geom Proto

Geom Proto

Geom Proto

Geom Proto

Geom Proto
}
\references{
\url{https://groups.google.com/forum/?fromgroups=#!topic/ggplot2/9cFWHaH1CPs}
}
\keyword{datasets}
\keyword{internal}
