% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalt-package.r
\docType{package}
\name{ggalt}
\alias{ggalt}
\alias{ggalt-package}
\title{Extra Geoms, Stats and Coords for 'ggplot2'}
\description{
A compendium of 'geoms', 'coords' and 'stats' for 'ggplot2' 2.0+, including
splines, 1d and 2d densities, univariate average shifted histograms and
a new map coordinate system based on 'PROJ.4'-based projections.
}
\author{
Bob Rudis (@hrbrmstr)
}

