% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggalt-package.r
\docType{package}
\name{ggalt}
\alias{ggalt}
\alias{ggalt-package}
\title{Extra Geoms, Stats and Coords for 'ggplot2'}
\description{
A package containing additional geoms, coords and stats for the revamped
(late 2015) version of ggplot2.
}
\author{
Bob Rudis (@hrbrmstr)
}

