% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalt-package.r
\docType{package}
\name{ggalt}
\alias{ggalt}
\alias{ggalt-package}
\title{Extra Geoms, Stats, Coords, Scales & Fonts for 'ggplot2'}
\description{
A package containing additional geoms, coords, stats, scales & fonts
for ggplot2 2.0+
}
\author{
Bob Rudis (@hrbrmstr)
}
