% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.r
\name{list_avatars}
\alias{list_avatars}
\title{List Pokemon avatar names}
\usage{
list_avatars()
}
\description{
List Pokemon avatar names
}
\note{
Warning: huge! list\cr
  \cr
  Pokémon & Pokémon character names are trademarks of Nintendo.
}
\references{
\href{http://www.pokegraphs.com/}{Original JSON color list}.
}

