% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.r
\name{pokemon_pal}
\alias{pokemon_pal}
\title{Pick a Pokemon palette}
\usage{
pokemon_pal(avatar = "bulbasaur")
}
\arguments{
\item{avatar}{avatar name. Use \code{list_avatars()} to see them all!}
}
\description{
Pick a Pokemon palette
}
\note{
Pokémon & Pokémon character names are trademarks of Nintendo.
}
\references{
\href{http://www.pokegraphs.com/}{Original JSON color list}.
}

